requirements.txt
streamlit
torch
torchvision
timm
opencv-python-headless
albumentations
pillow
numpy